set function as "private"
